use shared market utilities in uniswap provider
